+++ lib/src/analyzers/lint_analyzer/rules/rules_list/avoid_nested_conditional_expressions/config_parser.dart
@@ -1,0 +1,16 @@
+part of 'avoid_nested_conditional_expressions.dart';
+
+class _ConfigParser {
+  static const _acceptableLevelConfig = 'acceptable-level';
+
+  static const _defaultAcceptableLevel = 1;
+
+  static int parseAcceptableLevel(Map<String, Object> config) {
+    final level = _parseIntConfig(config, _acceptableLevelConfig);
+
+    return level != null && level > 0 ? level : _defaultAcceptableLevel;
+  }
+
+  static int? _parseIntConfig(Map<String, Object> config, String name) =>
+      config[name] != null ? int.tryParse(config[name].toString()) : null;
+}